Code refactor: separeted module builder, improved command line options handling.
