+++ .gitignore
@@ -1,3 +1,3 @@
-CMakeLists.txt.user
+CMakeLists.txt.user*
 external/*
 build/*
